feat: prevent attaching images to models without image input support (#171)
